fix: applying requested changes from PR review
