+++ frontend/components/auth/SignUp/SignUpOptionsForm.tsx
@@ -35,7 +35,7 @@ const SignUpOptionsForm: React.FC<SignUpOptionsFormProps> = ({
 			<Text size="md">
 				The email
 				<Text size="md" fontWeight="medium">
-					{` ${emailName} `}
+					{emailName}
 				</Text>
 				supports login with company SSO (Single Sign-on)
 			</Text>
